API/requirements.txt
django==2.1.2
django-rest-framework
psycopg2-binary
django-cors-headers
